Add files via upload
Who is The Next World Cup Winner?

Project using Data and Information Visualization to predict the next World Cup Winner of 2022. 
This project was done one year ago and Argentina just Won the World Cup. The Data Shows that Brazil was the favorite, Argentina is the second favorite followed by England. The Data always considered France not an option for winning.

- Use of Data
- Different Graphs and Charts
-Color, Size, Hue
-Storytelling technique